pyproject.toml
[tool.poetry]
name = "smit-test-project"
version = "0.1.0"
description = "This app is API service for quantity of price for safe service of isurance."
authors = ["Dmitry Malyshev <i@dmitrymalyshev.ru>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
alembic = "^1.14.0"
sqlalchemy = "^2.0.36"
fastapi = "^0.115.5"
pytest-dotenv = "^0.5.2"
pydantic-settings = "^2.6.1"
aiosqlite = "^0.20.0"
uvicorn = "^0.32.0"
asyncpg = "^0.30.0"
python-multipart = "^0.0.17"
pytest-asyncio = "^0.24.0"


[tool.poetry.group.dev.dependencies]
autopep8 = "^2.3.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
